Add support for request offers to update subscriber connections
The "sid" parameter of "requestOffer" messages is now checked to update
the current Janus connection when it matches the "sid" of the subscriber
rather than closing it and creating a new one.

Updating a connection is needed for clients to renegotiate their
received media without interrupting the current connection. For example,
to stop receiving video by triggering a new offer from Janus and then
setting the video transceiver direction as inactive before creating the
answer.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>