Remove "update" parameter in favour of "sid"
The "update" parameter was used to know if an offer should be treated by
the clients as a renegotiation or a new connection. However, since the
introduction of "sid" this is implicitly known (if the "sid" matches an
existing connection then it is an update, otherwise it is a new
connection), so that specific parameter is no longer needed.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>