Fix subscribers not closed when publisher is closed in Janus 1.x
In Janus 0.x when a publisher is closed the subscribers are also
automatically closed (if configured to do so, which is the default).
However, in Janus 1.x, as a subscriber can be connected to several
publishers, when a publisher is closed its subscribers just receive an
"updated" event with the new state of the streams.

As multistream connections are not used yet in Talk it still makes sense
to close its subscribers when a publisher is closed, so subscribers are
now automatically closed if they receive an "updated" event and all the
media streams are inactive (note that data channels will be always
active).

Note that, although it should not be received with the current usage of
Janus API, an "updated" event without streams could be received if the
subscriber was updated but there were no changes in the streams, so in
that case the subscriber should not be closed.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>